services/api/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.11.3
python-multipart==0.0.20
openai==1.76.2
pdfplumber==0.11.6
python-docx==1.1.2
openpyxl==3.1.5
tqdm==4.67.1
